requirements.txt
pandas==2.2.2
openpyxl==3.1.2
requests>=2.32.4
python-dotenv>=1.0.0
gunicorn==22.0.0
uvicorn
openai==1.99.9
google-adk>=1.10.0
litellm>=1.53
python-docx
google-cloud-vision>=3.7.0
google-cloud-firestore>=2.14.0
google-cloud-aiplatform>=1.38.0
Pillow>=10.0.0
google-auth==1.2.0